sched/tune: Rename dynamic_boost parameter to sched_boost
This was confusing to deal with given that it had the same name as the
Dynamic Schedtune Boost framework. It will be more apt to call it
sched_boost given that it was created to work with the sched_boost
feature in CAF devices.

The new tunable can be found in /dev/stune/*/schedtune.sched_boost

Signed-off-by: joshuous <joshuous@gmail.com>
(adapt for redmagic)
Signed-off-by: engstk <eng.stk@sapo.pt>

Signed-off-by: engstk <eng.stk@sapo.pt>